Moved the verticle and handler to existing consortium package
