+++ src/main/java/org/folio/rest/impl/InitApiImpl.java
@@ -11,7 +11,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.rest.resource.interfaces.InitAPI;
 import org.folio.services.caches.ConsortiumDataCache;
-import org.folio.services.consortia.ShadowInstanceSynchronizationVerticle;
+import org.folio.services.consortium.ShadowInstanceSynchronizationVerticle;
 import org.folio.services.migration.async.AsyncMigrationConsumerVerticle;
 
 public class InitApiImpl implements InitAPI {
